Ajustes sequelize product CRUD

+++ services/productService.js
@@ -1,5 +1,6 @@
 import productRepository from '../repositories/productRepository.js'
 import supplierRepository from '../repositories/supplierRepository.js'
+import saleRepository from '../repositories/saleRepository.js'
 
 async function createProduct(product) {
     if (await supplierRepository.getSupplier(product.supplier_id)) {
@@ -17,14 +18,20 @@ async function getProduct(id) {
 }
 
 async function updateProduct(product) {
-    if (await supplierRepository.getSupplier(product.supplier_id)) {
+    if (await supplierRepository.getSupplier(product.supplierId)) {
         return await productRepository.updateProduct(product)
     }
     throw new Error("O supplier_id informado não existe")
 }
 
 async function deleteProduct(id) {
-    return await productRepository.deleteProduct(id)
+    const sales = await saleRepository.getSalesByProductId(id)
+    if (sales) {
+        throw new Error('Não é possível excluir o produto pois possui vendas.')
+    } else {
+        await productRepository.deleteProduct(id)
+    }
+
 }
 
 export default {
